add undotree plugin and set open to F5
Add 'mbbill/undotree' plugin to display undo tree.
When dealing with undo/redo vim automatically keeps the undo log
as a tree. The object can be retrieved with :echo undotree() call
however it's not very useful in such a state.
This plugins adds a separate undotree split buffer accessible with
:UndotreeToggle.
Configuration maps this to F5 key